PWN-7023 - review fixes
