+++ app/src/main/java/org/p2p/wallet/swap/jupiter/repository/SwapTransactionRepository.kt
@@ -1,3 +1,9 @@
 package org.p2p.wallet.swap.jupiter.repository
 
-interface SwapTransactionRepository
+import org.p2p.solanaj.utils.crypto.Base64String
+import org.p2p.wallet.swap.jupiter.repository.model.SwapRoute
+import org.p2p.wallet.utils.Base58String
+
+interface SwapTransactionRepository {
+    suspend fun createSwapTransactionForRoute(route: SwapRoute, userPublicKey: Base58String): Base64String
+}
